feat: support handling element display callback in section controller
