+++ Airwallex/Payment/Sources/CollectionViewManager/SectionController.swift
@@ -59,6 +59,38 @@
     /// this method will be called in `CollectionViewContext.performUpdates(...)`
     /// this will be the place for you to update items in your section controller
     func updateItemsIfNecessary()
+    
+    /// Called just before a cell is displayed on the screen.
+    /// - Parameters:
+    ///   - cell: The `UICollectionViewCell` that is about to be displayed.
+    ///   - itemIdentifier: The unique identifier for the item.
+    ///   - indexPath: The `IndexPath` of the cell in the collection view.
+    func willDisplay(cell: UICollectionViewCell, itemIdentifier: ItemType, at indexPath: IndexPath)
+    
+    /// Called when a cell is removed from the screen (no longer visible).
+    /// - Parameters:
+    ///   - cell: The `UICollectionViewCell` that was displayed but is now being removed.
+    ///   - itemIdentifier: The unique identifier for the item.
+    ///   - indexPath: The `IndexPath` of the removed cell.
+    func didEndDisplaying(cell: UICollectionViewCell, itemIdentifier: ItemType, at indexPath: IndexPath)
+    
+    /// Called just before a supplementary view (e.g., header or footer) is displayed.
+    /// - Parameters:
+    ///   - supplementaryView: The `UICollectionReusableView` that is about to be displayed.
+    ///   - indexPath: The `IndexPath` of the supplementary view in the collection view.
+    func willDisplay(supplementaryView: UICollectionReusableView, at indexPath: IndexPath)
+    
+    /// Called when a supplementary view (e.g., header or footer) is removed from the screen.
+    /// - Parameters:
+    ///   - supplementaryView: The `UICollectionReusableView` that was displayed but is now being removed.
+    ///   - indexPath: The `IndexPath` of the removed supplementary view.
+    func didEndDisplaying(supplementaryView: UICollectionReusableView, at indexPath: IndexPath)
+    
+    /// Called just before the section managed by this section controller is displayed.
+    func sectionWillDisplay()
+    
+    /// Called when the section managed by this section controller is no longer visible.
+    func sectionDidEndDisplaying()
 }
 
 extension SectionController {
@@ -79,6 +111,30 @@ extension SectionController {
     func anySectionController() -> AnySectionController<SectionType, ItemType> {
         AnySectionController(self)
     }
+    
+    func willDisplay(cell: UICollectionViewCell, itemIdentifier: ItemType, at indexPath: IndexPath) {
+        // do nothing by default
+    }
+    
+    func didEndDisplaying(cell: UICollectionViewCell, itemIdentifier: ItemType, at indexPath: IndexPath) {
+        // do nothing by default
+    }
+    
+    func willDisplay(supplementaryView: UICollectionReusableView, at indexPath: IndexPath) {
+        // do nothing by default
+    }
+    
+    func didEndDisplaying(supplementaryView: UICollectionReusableView, at indexPath: IndexPath) {
+        // do nothing by default
+    }
+    
+    func sectionWillDisplay() {
+        // do nothing by default
+    }
+    
+    func sectionDidEndDisplaying() {
+        // do nothing by default
+    }
 }
 
 /// Type erasor for SectionController
@@ -93,6 +149,12 @@ class AnySectionController<SectionType: Hashable & Sendable, ItemType: Hashable 
     private let _context: () -> CollectionViewContext<SectionType, ItemType>
     private let _bindContext: (CollectionViewContext<SectionType, ItemType>) -> Void
     private let _prepareItemUpdates: () -> Void
+    private let _willDisplayCell: (UICollectionViewCell, ItemType, IndexPath) -> Void
+    private let _didEndDisplayingCell: (UICollectionViewCell, ItemType, IndexPath) -> Void
+    private let _willDisplaySupplementaryView: (UICollectionReusableView, IndexPath) -> Void
+    private let _didEndDisplayingSupplementaryView: (UICollectionReusableView, IndexPath) -> Void
+    private let _sectionWillDisplay: () -> Void
+    private let _sectionDidEndDisplaying: () -> Void
     
     var context: CollectionViewContext<SectionType, ItemType>! { _context() }
     var section: SectionType { _section() }
@@ -108,6 +170,12 @@ class AnySectionController<SectionType: Hashable & Sendable, ItemType: Hashable 
         self._section = { sectionController.section }
         self._context = { sectionController.context }
         self._prepareItemUpdates = { sectionController.updateItemsIfNecessary() }
+        self._willDisplayCell = sectionController.willDisplay(cell:itemIdentifier:at:)
+        self._didEndDisplayingCell = sectionController.didEndDisplaying(cell:itemIdentifier:at:)
+        self._willDisplaySupplementaryView = sectionController.willDisplay(supplementaryView:at:)
+        self._didEndDisplayingSupplementaryView = sectionController.didEndDisplaying(supplementaryView:at:)
+        self._sectionWillDisplay = sectionController.sectionWillDisplay
+        self._sectionDidEndDisplaying = sectionController.sectionDidEndDisplaying
     }
     
     func bind(context: CollectionViewContext<SectionType, ItemType>) {
@@ -132,5 +200,29 @@ class AnySectionController<SectionType: Hashable & Sendable, ItemType: Hashable 
     
     func updateItemsIfNecessary() {
         _prepareItemUpdates()
+    }
+    
+    func willDisplay(cell: UICollectionViewCell, itemIdentifier: ItemType, at indexPath: IndexPath) {
+        _willDisplayCell(cell, itemIdentifier, indexPath)
+    }
+    
+    func didEndDisplaying(cell: UICollectionViewCell, itemIdentifier: ItemType, at indexPath: IndexPath) {
+        _didEndDisplayingCell(cell, itemIdentifier, indexPath)
+    }
+    
+    func willDisplay(supplementaryView: UICollectionReusableView, at indexPath: IndexPath) {
+        _willDisplaySupplementaryView(supplementaryView, indexPath)
+    }
+    
+    func didEndDisplaying(supplementaryView: UICollectionReusableView, at indexPath: IndexPath) {
+        _didEndDisplayingSupplementaryView(supplementaryView, indexPath)
+    }
+    
+    func sectionWillDisplay() {
+        _sectionWillDisplay()
+    }
+    
+    func sectionDidEndDisplaying() {
+        _sectionDidEndDisplaying()
     }
 }
